Add support for points in user story operations
Added `Points? points` parameter to `CreateAsync` and `UpdateAsync` methods in `UserStoryTool.cs` to allow specifying effort estimates for UX, Design, Front-end, and Back-end. Updated logic to handle the `points` parameter by including it in the `data` dictionary when provided. Also added a reference to the `Taiga.Api.Models` namespace to support the `Points` type.